requirements.txt
ultralytics==8.0.121
scikit-learn==0.21.0
tensorflow==2.11.0
scikit-image==0.19.3
filterpy==1.4.5
cvzone==1.5.6
opencv-python~=4.7.0.72
numpy~=1.21.6
scipy~=1.7.3
mysql-connector-python==8.0.33

pandas~=1.3.5
adtk~=0.6.2
matplotlib~=3.5.3
plotly~=5.15.0
Flask~=2.2.5
Flask-SocketIO~=5.3.5